+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 - **New Data Source:** `netapp-ontap_autosupport` ([#555](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/555))
 - **New Resource:** `netapp-ontap_autosupport` ([#555](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/555))
 
+FEATURES:
+
+- **New Data Source:** `netapp-ontap_s3_group` ([#591](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/591))
+- **New Data Source:** `netapp-ontap_s3_groups` ([#591](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/591))
+- **New Resource:** `netapp-ontap_s3_group` ([#592](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/592))
+
 # 2.3.0 (2025-07-22)
 
 DOCUMENTATION: